Update qAndora.py
Moved the show/hide toggle in the system tray icon menu to the top and added an icon to it. And moved "self.radioPlayer.setChangeCallBack(self.songChangeQ)" to it's own function. "self.stationBox.activated[str].connect(self.stationChange)" should stay in assignWidgets since it's a widget. But "self.radioPlayer.setChangeCallBack(self.songChangeQ)" doesn't belong there. Also got rid of a variable I introduced that was not needed.